app/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
asyncpg==0.29.0
redis==5.2.1
pydantic==2.9.2
pydantic-settings==2.5.2
PyJWT==2.10.1
python-multipart==0.0.9
python-dotenv==1.0.1
alembic==1.15.1
sqlalchemy[asyncio]==2.0.36
psycopg[pool]==3.2.3
psycopg2-binary

